[REF] mail: mail.chatter service
