+++ addons/mail/static/src/new/activity/activity_list_popover_item.js
@@ -36,7 +36,7 @@ export class ActivityListPopoverItem extends Component {
         this.state = useState({ hasMarkDoneView: false });
         if (this.props.activity.activity_category === "upload_file") {
             this.attachmentUploader = useAttachmentUploader(
-                this.env.services["mail.messaging"].getChatterThread(
+                this.env.services["mail.chatter"].getThread(
                     this.props.activity.res_model,
                     this.props.activity.res_id
                 )
