+++ addons/mail/static/src/new/activity/activity_markasdone_popover.js
@@ -18,6 +18,7 @@ export class ActivityMarkAsDone extends Component {
     setup() {
         this.messaging = useMessaging();
         this.threadService = useState(useService("mail.thread"));
+        this.chatter = useState(useService("mail.chatter"));
         this.textArea = useRef("textarea");
         onMounted(() => {
             this.textArea.el.focus();
@@ -33,7 +34,7 @@ export class ActivityMarkAsDone extends Component {
 
     async onClickDone() {
         const { res_id: resId, res_model: resModel } = this.props.activity;
-        const thread = this.messaging.getChatterThread(resModel, resId);
+        const thread = this.chatter.getThread(resModel, resId);
         await this.env.services["mail.activity"].markAsDone(this.props.activity);
         if (this.props.reload) {
             this.props.reload(this.props.activity.res_id, ["activities"]);
@@ -43,7 +44,7 @@ export class ActivityMarkAsDone extends Component {
 
     async onClickDoneAndScheduleNext() {
         const { res_id: resId, res_model: resModel } = this.props.activity;
-        const thread = this.messaging.getChatterThread(resModel, resId);
+        const thread = this.chatter.getThread(resModel, resId);
         if (this.props.onClickDoneAndScheduleNext) {
             this.props.onClickDoneAndScheduleNext();
         }
