+++ addons/mail/static/src/new/core/messaging.js
@@ -643,56 +643,6 @@ export class Messaging {
     // actions that can be performed on the messaging system
     // -------------------------------------------------------------------------
 
-    getChatterThread(resModel, resId) {
-        const localId = createLocalId(resModel, resId);
-        if (localId in this.state.threads) {
-            if (resId === false) {
-                return this.state.threads[localId];
-            }
-            // to force a reload
-            this.state.threads[localId].status = "new";
-        }
-        const thread = Thread.insert(this.state, {
-            id: resId,
-            model: resModel,
-            type: "chatter",
-        });
-        if (resId === false) {
-            const tmpId = `virtual${this.nextId++}`;
-            const tmpData = {
-                id: tmpId,
-                author: { id: this.state.user.partnerId },
-                body: _t("Creating a new record..."),
-                message_type: "notification",
-                trackingValues: [],
-            };
-            Message.insert(this.state, tmpData, thread);
-        }
-        return thread;
-    }
-
-    async fetchChatterData(
-        resId,
-        resModel,
-        requestList = ["activities", "followers", "attachments", "messages"]
-    ) {
-        if (requestList.includes("messages")) {
-            this.thread.fetchNewMessages(Thread.insert(this.state, { model: resModel, id: resId }));
-        }
-        const result = await this.rpc("/mail/thread/data", {
-            request_list: requestList,
-            thread_id: resId,
-            thread_model: resModel,
-        });
-        if ("attachments" in result) {
-            result["attachments"] = result["attachments"].map((attachment) => ({
-                ...attachment,
-                originThread: Thread.insert(this.state, attachment.originThread[0][1]),
-            }));
-        }
-        return result;
-    }
-
     fetchPreviews = memoize(async () => {
         const ids = [];
         for (const thread of Object.values(this.state.threads)) {
@@ -761,16 +711,5 @@ export class Messaging {
         return this.rpc("/mail/attachment/delete", {
             attachment_id: attachment.id,
         });
-    }
-
-    /**
-     * @param {import("@mail/new/core/follower_model").Follower} follower
-     */
-    async removeFollower(follower) {
-        await this.orm.call(follower.followedThread.model, "message_unsubscribe", [
-            [follower.followedThread.id],
-            [follower.partner.id],
-        ]);
-        follower.delete();
     }
 }
